Moving settings to JSON WIP

+++ spotify-mini-player/callback.php
@@ -10,52 +10,14 @@ $w = new Workflows('com.vdesabou.spotify.mini.player');
 
 
 //
-// Read settings from DB
+// Read settings from JSON
 //
-$getSettings = 'select oauth_client_id,oauth_client_secret,oauth_redirect_uri from settings';
-$dbfile = $w->data() . '/settings.db';
 
-try {
-	$dbsettings = new PDO("sqlite:$dbfile", "", "", array(PDO::ATTR_PERSISTENT => true));
-	$dbsettings->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-	$dbsettings->query("PRAGMA synchronous = OFF");
-	$dbsettings->query("PRAGMA journal_mode = OFF");
-	$dbsettings->query("PRAGMA temp_store = MEMORY");
-	$dbsettings->query("PRAGMA count_changes = OFF");
-	$dbsettings->query("PRAGMA PAGE_SIZE = 4096");
-	$dbsettings->query("PRAGMA default_cache_size=700000");
-	$dbsettings->query("PRAGMA cache_size=700000");
-	$dbsettings->query("PRAGMA compile_options");
-} catch (PDOException $e) {
-	displayNotificationWithArtwork("Error[callback.php]: cannot set PDO settings",'./images/warning.png', 'Error!');
-	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
-	$dbsettings=null;
-	return;
-}
+$settings = getSettings($w);
 
-try {
-	$stmt = $dbsettings->prepare($getSettings);
-	$settings = $stmt->execute();
-
-} catch (PDOException $e) {
-	displayNotificationWithArtwork("Error[callback.php]: cannot prepare settings",'./images/warning.png', 'Error!');
-	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
-	$dbsettings=null;
-	return;
-}
-
-try {
-	$setting = $stmt->fetch();
-}
-catch (PDOException $e) {
-	displayNotificationWithArtwork("Error[callback.php]: cannot fetch settings",'./images/warning.png', 'Error!');
-	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
-	return;
-}
-
-$oauth_client_id = $setting[0];
-$oauth_client_secret = $setting[1];
-$oauth_redirect_uri = $setting[2];
+$oauth_client_id = $settings->oauth_client_id;
+$oauth_client_secret = $settings->oauth_client_secret;
+$oauth_redirect_uri = $settings->oauth_redirect_uri;
 
 try {
 	$session = new SpotifyWebAPI\Session($oauth_client_id, $oauth_client_secret, $oauth_redirect_uri);
@@ -71,23 +33,47 @@ try {
 			$api->setAccessToken($session->getAccessToken());
 			$user = $api->me();
 
-			$updateSettings = "update settings set oauth_access_token=:oauth_access_token,oauth_expires=:oauth_expires,oauth_refresh_token=:oauth_refresh_token,country_code=:country_code,display_name=:display_name,userid=:userid";
-			try {
-				$stmt = $dbsettings->prepare($updateSettings);
-				$stmt->bindValue(':oauth_access_token', $session->getAccessToken());
-				$stmt->bindValue(':oauth_expires', time());
-				$stmt->bindValue(':oauth_refresh_token', $session->getRefreshToken());
-				$stmt->bindValue(':country_code', $user->country);
-				$stmt->bindValue(':display_name', $user->display_name);
-				$stmt->bindValue(':userid', $user->id);
-				$stmt->execute();
+		    $ret = updateSetting($w,'oauth_access_token',$session->getAccessToken());
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
 
-			} catch (PDOException $e) {
-				handleDbIssuePdoEcho($dbsettings,$w);
-				$dbsettings=null;;
-				exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
-				return;
-			}
+		    $ret = updateSetting($w,'oauth_expires',time());
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
+
+		    $ret = updateSetting($w,'oauth_refresh_token',$session->getRefreshToken());
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
+
+		    $ret = updateSetting($w,'country_code',$user->country);
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
+
+		    $ret = updateSetting($w,'display_name',$user->display_name);
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
+
+		    $ret = updateSetting($w,'userid',$user->id);
+		    if($ret == false) {
+			 	echo "There was an error when updating settings";
+			 	exec("kill -9 $(ps -efx | grep \"php -S localhost:15298\"  | grep -v grep | awk '{print $2}')");
+			 	return;
+		    }
 
 			echo "Hello $user->display_name ! You are now successfully logged and you can close this window.";
